Update highscores.py to move to is_high_score()
Is_high_score() returns True or False to the caller to let them know if the score is really a new high score.  Often the caller wants to notify the user of their achievement and congratulate them.  We also now check the initial score to ensure that it is numeric.